\name{teeth4}
\alias{teeth4}
\title{A simulated data set to demonstrate the `baymvb' package}
\description{
   `teeth4' data is a computer simulated multivariate binary data
   	set to demonstrate the `baymvb' package
}

\keyword{internal}

 \author{S. M. Mwalili \email{samuel\_mwalili@yahoo.com}}
