\name{xpndCor}
\alias{xpndCor}	

\title{The correlation matrix from the posterior samples}
\description{
  The correlation matrix from the posterior samples  of 'baymvb' objects
}
\usage{
  	xpndCor(obj)
}


\arguments{
  \item{obj}{Posterior samples from `baymvb'}  
}

\value{
  Returns the correlation matrix from the posterior samples  of 'baymvb' objects
}

\seealso{
\code{\link{densPlot}},\code{\link{baymvb}}.
} 


\keyword{internal}

\author{S. M. Mwalili \email{samuel\_mwalili@yahoo.com}}

